+++ src/MetaheuristicsPlatform/Algorithms/CuckooSearch/CuckooSearchReferences.cs
@@ -1,0 +1,14 @@
+using MetaheuristicsPlatform.Classification;
+
+namespace MetaheuristicsPlatform.Algorithms.CuckooSearch;
+
+public static class CuckooSearchReferences
+{
+    public static ScientificReference YangDeb2009 { get; } =
+        new(
+            "Yang & Deb",
+            2009,
+            "Cuckoo Search via Levy Flights",
+            "2009 World Congress on Nature & Biologically Inspired Computing (NaBIC), 210-214",
+            "10.1109/NABIC.2009.5393690");
+}